+++ src/core/shared/enums/vehicle.ts
@@ -45,7 +45,9 @@ export const VEHICLE_EVENTS = {
     SET_LOCK: 'Vehicle-Set-Lock',
     SET_DOOR: 'Vehicle-Set-Door',
     SET_ENGINE: 'Vehicle-Set-Engine',
-    SET_SEATBELT: 'Vehicle-Seatbelt'
+    SET_SEATBELT: 'Vehicle-Seatbelt',
+    PUSH: 'Vehicle-Push',
+    STOP_PUSH: 'Vehicle-Stop-Push'
 };
 
 export const VEHICLE_DOOR_STATE = {